# openApi/a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'Ask_LEIC Web API'
  description: 'Web Resources Specification (A7) for ask_LEIC'
servers:
- url: http://localhost:8001/ #http://lbaw.fe.up.pt #MUDAR
  description: Production server
...
paths:
  /login:
    get:
      operationId: R101
      summary: 'R101: Login Form'
      description: 'Provide the login form. Access: PUB'
      tags:
        - 'M01: Authentication and Individual Profile'
      responses:
        '200':
          description: 'Login form displayed successfully.'
        '500':
          description: 'Internal server error while loading the login form.'

    post:
      operationId: R102
      summary: 'R102: Login Action'
      description: 'Processes the login form submission and authenticates the user. Access: PUB'
      tags:
        - 'M01: Authentication and Individual Profile'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: 'The user’s email address.'
                password:
                  type: string
                  format: password
                  description: 'The user’s password.'
              required:
                - email
                - password
      responses:
        '302':
          description: 'Successful authentication. Redirect to user profile.'
          headers:
            Location:
              schema:
                type: string
              examples:
                RedirectToProfile:
                  description: 'Redirect to the authenticated user’s profile.'
                  value: '/users/{id}'
        '400':
          description: 'Bad request (e.g., missing or invalid credentials).'
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Invalid email format or missing parameters.'
        '401':
          description: 'Unauthorized. Invalid email or password.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Authentication failed. Please check your credentials.'
        '403':
          description: 'Forbidden. User account is disabled or banned.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'Your account has been disabled. Please contact support.'
        '500':
          description: 'Internal server error during login attempt.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: 'An unexpected error occurred. Please try again later.'
...